requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
decorator==4.1.2
dj-database-url==0.4.2
Django==1.9.13
future==0.16.0
geocoder==1.33.0
gunicorn==19.7.1
idna==2.6
psycopg2==2.7.3.2
PyMsgBox==1.0.6
pyserial==3.4
pytz==2017.2
ratelim==0.1.6
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
whitenoise==3.3.1
